feat: Update BuildLinux.sh to support USE_SYSTEM_DEPS
Updated BuildLinux.sh to include a new command-line option  which enables the  CMake option. When  is used, the script will configure ElegooSlicer to use system-installed shared libraries and will not force static linking (). This allows users to leverage system-level dependencies for a more integrated build on Arch Linux and similar distributions.